Beat my head against crazy timing issues
Unit tests that work sporadically and fail without explanation are
incredibly frustrating.

In this case, it was totally worthwhile.

My scheme to use the same async socket for reading/writing messages
from/to the 0mq side was doomed. Sometimes the async loop was reading
its own messages before the main thread could pick it back up.