Refining unit test to comm from in to out
It really looks like there's a very odd interaction happening with the
Components. Stopping is failing. Then, when it starts for the next
test (with what should be a completely different System and set of
socket addresses), it's receiving a message from outside (how? I don't
have any idea). It looks like that gets forward somewhere, then the next
message (the first for this test run) I send seems to get swallowed.

There's also still a screwiness with the way messages I'm
sending/receiving manually (which still works) wind up being serialized
and transmitted (e.g. a byte array winds up encoded as "\"\\\\W\"")

Actually, it looks like, in my last test, that was just "\\W". I'm using
whidbey, so it looks as if it's just sending the first character of its
encoded version (I think it should be "#whidbey/bin [...]", so that
theory may be garbage).