Switch some async channels to Components
So I can feed them as dependencies through a System and not create/stop
them at the reader. Still isn't perfect (since the writer should really
be handling cleanup), but I think it's much more idiomatic.

Got severel unit tests working with this change, though several others
are broken. Haven't had a chance to track down whether that's because of
this change or due to something seemingly unrelated elsewhere.