Start converting async-zmq from schema to spec
Right now they're playing nicely side by side, so it's only about half done.

This involves a lot of other tendrils that spread out to quite a few
other namespaces.

Almost as a side-effect, discovered that my scheme for building an event
loop as its own SystemMap and have clients use that doesn't really
work. The Context dependency wasn't getting fed to the Socket correctly.

So converted that function to return the component-dsl description of
those pieces to see if/hope that this approach works better.