Working on broken spec (et al)
Started w/ an undecipherable error about :event-loopless-pair, when I
started checking it in the pre-condition of run-async-loop (something
I'd flagged w/ TODO: ^:always-validate when I converted it from
prismatic schema).

(The problem seems to be in cljeromq's spec that wraps ZMQ$Socket).

Added a broken unit test for it.

Got rid of the obsolete build-event-loop-broken function.

Fixed some whitespace problems.