Narrow down a problem w/ Server startup
Not necessarily the problem, and almost definitely not the really
interesting part of it.

But it really looks as though my socket constructor is getting called
with both namespaced and non-namespace keywords. Which almost definitely
violates the spec.

Not sure how that bleeds over into the generator, but it's the most
likely candidate I've found so far.