Start async-loop before its 0mq counterpart
This way, if it throws an exception (such as because the input
parameters don't satisfy its spec), and I'm working with a :bind socket,
I won't have to restart the JVM just to release that socket.

Also, relax spec checking on run-async-loop! parameters.

Actually, just skip it completely to try to get Frereth Server
starting (which seems to work).

I think I see the actual problem, but this was an important step.