Misplaced paren was throwing even if spec matched
Which it doesn't. Well, something is wrong with it. I get errors about
the generated object for the reader handlers not supporting .recv, which
is far enough from the truth to make me think I have another error in
something different that looks like it's coming from here.

Problem w/ previous commit really was the namespace aliasing.