Cobble together aleph router unit tests (WIP)
The main change is really in the server event loop. Don't automatically
deref the take!

That part seems like a pretty solid fix for the issues that have plagued
me for the past couple of days where trying to write back to the socket
simply wasn't showing up back at the client.

The cobbled part is the way I'm synchronizing put! and take! in the unit
test.

Right now, I'm just throwing in long enough Thread/sleep calls to make
it work. This approach is obviously wrong.

But it's forward progress.