Start translating actual CurveCP messages
This meant bringing aleph back into the picture, since it has a
perfectly fine UDP implementation that would be silly to reinvent.

And I'm starting to pick apart the cookie response that the client will
receive in response to its 'HELLO' message.

This meant starting to string together the actual messages that will be
going back and forth and defining the actual packets with a binary
protocol definer.