WIP client: Initiate translated, theoretically
The part that pulls the Messages from the "child" process is pretty
rough, if only because I haven't had time to dig into what it's actually
doing to decide how much translating it needs.

And I definitely have unanswered questions.

But it's up to the point where I can at least start testing it, once
I've made the same progress on the server side.