broken wip: chasing CP server start/stop tail
It does start now. And then I get an NPE from java SecureRandom when I
try to stop.

And *then* I get obnoxious compilation errors from Mount when I try to
recompile my fixes.

I think that's probably because it's trying to helpfully stop whatever I
have running now so it doesn't leak memory/resources after the refresh
orphans it. But this is a change that's definitely going to take some
getting used to.

Especially since the related error message is useless.