Start breaking Client handshake into init steps
There seem to be distinct mutable vs. immutable pieces here. And then
there are the gray areas like the byte arrays where the real bytes get
shuffled around.

I don't think the Client is a Component after all. It's more like a data
struct that happens to include some mutable state that should be one
member of a collection Component.

Checking in here/now as a milestone in case I decide later that I've
made a terrible mistake.