Dabble w/ buffy for building packets
This seems to be another way to not do what I want. They *do* have
support for supplying something like a nio ByteBuffer or netty ByteBuf,
but that seems to be strictly read-only.

Oh well. At least I have enough insight to how it works to accept that
it really wouldn't add enough to justify its inclusion.