Test my ByteBuf creation changes
They seem to work correctly.

But the test fails at the next step, when the client's trying to pull
the cookie that a server should be sending back (it really looks like
it's getting its HELLO packet again, especially since there shouldn't be
anything else it possibly could be receiving).