Isolate latest problems to race conditions
Over the shared packet ByteBuf that I'm using for
a) storing bytes while I'm setting them up to send/recv
b) actually sending off to the server

Seems fairly obvious that I really need a pool of direct ByteBuf
instances to use for i/o and a local ByteBuf for doing the actual bit
twiddling.

Committing this version because "fairly obvious" is so frequently
flat-out wrong.