Get curve client Hello Handshake test passing
At least, it looks like it's generating a message that ought to be a
valid "Hello" packet and sending it to the stream that will
theoretically be connected to the server.

And then the associated Agent switches to the appropriate error state
when I close the other half of that stream since mocking up the server
side is far beyond the scope of this test.

At least for now.