Push handshake test forward (WIP)
I'm getting the Hello packet from the client and sending it to the
server. But the server's acting as though I'm closing the channel
instead of writing the packet.

The basic outline for the rest of the test should be mostly correct,
but I haven't managed to get past this point to verify.