Debugging curve handshake failure
Was trying to randomize a direct ByteBuf as though it were a plain
Byte Array.

Moved each link in the chain of Deferreds that forms the CurveCP
handshake into its own function to try to isolate the latest failure.

Which turns out to be happening just before I start that running that
chain.