Get CurveCP handshake test up to server receive
Well, it was up to that point and failing earlier. Because I was
throwing a NotImplementedException.

Now it's failing because that method isn't even trying to do anything,
so the client times out while it's waiting for the server response.

At least now I'm back to the point that I can see where it's breaking.