WIP: Debugging Curve Server HELLO packet decrypt
Add a little extra debugging to TweetNaclFast to try to pinpoint the
failures. I'm not sure how to reload that (or even if it's possible)
without restarting my REPL session. So I went ahead and translated what
it does to a couple of routines in shared.

They're just straight-up failing on the decryption call.