Rein in my over-optimism
Got ahead of myself w/ previous commit message.

My handshake test didn't magically start passing. I accidentally tested
something that was already working.

It was getting further. That version was able to decode the HELLO packet
when it arrived at the server from the client.

But the server wasn't able to build the cookie in response.

This version builds the cookie and sends it back, but it looks like it
turns around immediately and pulls it back off its incoming channel.