Refactor the byte-streams ns alias in server
Make it match the one I chose for client. It seems less likely to cause
confusion.

This brought my attention to the other major problem that I solved in
the previous commit: I thought I was calling a function for
side-effects, but it was returning the value it was calculating, and
then that was getting silently discarded.