Add place for client to buffer child messages
And a "constants" function for checking the limit, along with
appropriate unit tests.

There was also a bit of cleaning up dead code.

This version looks like the handshake test passes at the REPL, but
that's mostly because the server rejects the Initiate packet because
it's still missing any Message bytes.