Fix couple of silly problems with interaction-test
1. One test was relying on a debugging change I'd made to TweetNaclFast
that exposed the shared key as a public property. I just had to
eliminate the test that was comparing this against the values I was
calculating.

2. Handshake test had a misplaced closing parenthesis when it tried to
"fork" the child "process" that was causing an ArityException.

This takes me back to the basic problem I was having last time where
that test is creating a false positive. But now I'm sure the problem is
in my code.

And I rolled back that "welcome" message in dev/user.clj. It was showing
up every time I refreshed that ns, which was just annoying.