More debugging recent refactorings
Server handles HELLO correctly with a Cookie response packet.

Then the client is failing to produce a Vouch because its stream from
the mocked-up child is closed before it gets a chance to read from it.

I don't have anything that even vaguely resembles useful timestamps in
the logs, but I can see where the mock child fails to send its payload.