Deal with recent refactoring fallout: client side
More or less.

Part of problem that I called out in comment to previous commit was that
I needed to add a timeout to the mock-child's deref before checkin its
try-put success.

That revealed a slew of NPEs from changes to nonsensical keys.

It also pointed out more nonsensical keys, but I'm not tackling those
just yet.

It looks as though the client is sending the Initiate packet back to the
server again, but the message just disappears.