Get through Initiate packet NPE logjam
The trick was tracking down the actual problem.

Root cause was an incomplete refactoring of the constant's template
description. Extracting that packet on the server failed because it was
trying to modify the length field under ::vouch instead of
::vouch-wrapper.

Still don't have any idea why the stack trace was coming back empty.

This version gets a couple of steps further before it fails.

But now I think I'm getting back into the area of debugging fresh code
instead of something that I recently broke.