Convert my crypto/open-box wrapper: return ByteBuf
Returning a subvec meant that I had to convert it to a ByteBuf before I
could decompose it, and I don't foresee a scenario where I wouldn't want
to do that.

So cut out the middleman, since I already have a Byte array that I just
decrypted to conveniently slice.