Refactor client into multiple pieces
I'm not particularly fond of the way this split wound up working: I just
grabbed pieces and tried to more-or-less mirror what I've done to split
up the server side.

Far too much wound up in the state ns, but that's a bigger problem for
some other time.

For now, I'm just trying to figure out why the client's copy of the
server's long-term public key doesn't seem to match anything
else (especially when it rolls around to having the server decrypt the
inner-most portion of the Initiate packet's Vouch).