Wrapping up Initiate handling
The Server side is "forking" a child "process", and I have a message
stream set up between the two.

I have a unit test that should be sending the Client's initial message
to the Server's child, which should echo it back.

That part isn't working, but I think that's because it bleeds into the
Server's Message handling. I haven't had a chance to take that beyond a
preliminary "I think this is roughly what the entry to this piece ought
to look like" phase.