Work on disagreement: ->child vs receivebuf (WIP)
I wrote about half of the from-parent message handler expecting to just
conj a ByteBuf onto a vector of incoming message blocks. And the other
half expecting to feed its bytes into a circular receivebuf the way the
reference implementation does.

I still have at least one piece that still needs to be reconciled, but
this was a big chunk of the reconciliation between the two.