Debugging send-from-parent echo unit test (WIP)
2 major, subtle changes in here:

1. Realize that the "data block size/EOF flags" short is really based
around the length of the message block so we know when the zero padding
ends
2. Uncomment a sanity check that verifies that we have as many bytes
available as the header proclaims.

This version doesn't work, because the numbers aren't matching up.

It's also an important commit, because I think this is enough reason to
just switch to decoding everything in one fell swoop instead of
piece-meal the way I'm trying to do now (because that matches the
reference implementation).

So I want a milestone to roll back to in case the pending departure is a
big failure.