Switch to using last night's message deserializer
Mostly.

Making this changed fixed 1 serious bug where sometimes I was pulling
the front end of the child-buffer queue and other times the back, in
order to read the next set of bytes.

This at least gets me back to where I was before I decided the previous
implementation was just too complex to be worth trying to save.