Post-deserialization message coping
The code that extracts the actual message body from its packet still
needed to be tweaked with the fact that I've already read the bytes from
the packet's ByteBuf.

After that returns, I wasn't calling the ack-prep function to build the
message that we send back to let the other side know we received it.

My echo unit test still fails with the same error I had before I made
these adjustments (that I'm sending a gibberish message), but these were
important steps.