Make a little more progress on echo test
Work around my dilemma about the circular dependency between the child
callback and the actual state agent by sticking the agent inside an
atom. It seems pretty awful, but the test is fighting the design.

Wound up moving the spec for ::state-agent out of message and into specs
ns so to-child can reference it. It turns out the access didn't make any
sense, but the rearrangement seems logical.

This version seems to get past the child callback and then go completely
off the rails by triggering the from-parent signal again, before
complaining about an illegal 0-length message and throwing a NPE with no
stack trace.