Fix problem with previous commit
Headers (etc) were getting added back in because I was trying to avoid
the overhead of copying a byte array by accessing the backing array
directly. So of course it ignored my carefully chosen reader/writer
indexes.

This version gets past that to fail when it tries to send the echo.

Except that it's failing in from-parent, which seems suspicious. Or
maybe it just looks like something weird along those lines because of
something unfamiliar (to me) about the way agents work.