Sort through in previous commit's problems (WIP)
...at least the next layer of them.

When I triggered the I/O loop, I forgot to account for the fact that
processing the incoming from-parent messages returns nil when there
aren't any.

So I was calling the final ->child callback with that.

Fixed some type hints where I changed from ByteBuf parameters to byte
arrays (along with related function/method transitions).

Lots of debug logging and comment improvements.

And I'm still sending the wrong thing to the ->child callback. But it's
quite a bit closer.