Fix some basic problems w/ previous commit (WIP)
The main problem was that I was prematurely popping my most recently
sent new message off the queue, so it wasn't there later when I tried to
move it to the queue of messages that have not yet been acknowledged.

But, after that, I think I got a parenthesis misaligned. So
deserializing was returning nil any time there was no initial padding to
skip. It seems like this should never have mattered, since there
should *always* be initial padding, but it caused problems.

I'm still getting gibberish at the server. After it sets up the ACK
flags, the message length is garbage. But this is the first real forward
progress I've made in days.