Get echo test working again (WIP)
Or maybe for the first time. The way I was handling gap acknowledgments
before was totally wrong.

The main key to fixing that was to track how many contiguous bytes have
been received. I think that's actually what receivebytes is for in the
original, but I didn't realize/understand that until I'd reinvented that
wheel. Oh well.

I haven't done much other testing, but I know I've broken
from-child-test.

I also realized that sticking a log wrapper in utils was futile: it
hides the actual ns where the logging is happening. So started backing
that out and just switching to a boilerplate function that sets up a log prefix.