Fix the easy "other" tests (WIP)
There are still that are broken. One of which has been that way for
months now. So I can't blame current work in this branch.

Right now, I'm more interested in the top-level message tests. They do
cover the consolidation that's breaking in the other that's having
issues, so that seems like a problem with the test.