Base sending on queue pointer, instead of nil
Original checks for "should I send from this queue next?" returned nil
for the state, if it wasn't appropriate to send.

Just don't set the field that holds the pointer to that queue instead.

This may be a performance hit (since I have to do a nested map
destructuring/lookup), but it seems cleaner/safer.