Refactor some other really annoying bits (WIP)
Like a gigantic let block that was modifying state very slightly, one
form at a time, to cope with an incoming ACK.

This version gets further in my "big block send" unit test than the past
few/several commits. It breaks when it tries to build message 4, because
that involves writing to a ByteBuf that has a 0 refCnt.

I have a hypothesis about how to fix that.

It also fails because I'm calling trigger-from-parent with something
that doesn't have a ->child key.

I strongly suspect the latter is really caused by a combination of the
former and a lack of error checking.