Fix a problem checking for block to send (WIP)
Had a misplaced if. I was trying to send a block before I'd checked to
see whether there was one to send.

This fixes that, but the "big send" test goes into a fast infinite loop
near the end, spinning with nothing to send. Haven't had a chance to dig
into that angle yet.