Combing through logs (WIP)
Adding comments and log improvements. With a couple of minor code tweaks
that shouldn't amount to anything.

I'm not sure this even compiles, but I don't think any of the changes
could have broken anything.

(I know, this is awful...I'm preserving my CIDER buffer because that's
where my logs are).