Try rearranging delays in big send unit test
The last failure I was seeing was that the final ACK wasn't getting back
to the client child before I deref'd the associated agent. So its final
address was wrong (even given the fundamental brokenness of the current
implementation).

I had a theory that I could move up the cancel trigger. Send that
message, do an await, and everything would hopefully sort itself out by
the time I got around to calling deref.

This didn't pan out. Now the race condition's far more obvious. But it
was worth the attempt.