Reduce state-wrapper parameters in message (WIP)
It wasn't needed in most of the places where I was using it.

Also got pickier about the wrapper function for building the log prefix
when I have that instead of the plain state value.

This is really just an incremental refactoring checkin. It compiles, but
I haven't run any tests yet.