Poke through broken echo test (WIP)
So far, I've really just managed to break it worse. Or, rather, make the
broken spots a lot more obvious.

Added a reference to test.check so I can actually try to validate the
message state spec.

And started throwing exceptions in the places where I obviously need to
make changes, but just am not willing to tackle at the moment.