Update echo unit test for passing byte-arrays...
instead of ByteBuf, to the client/server encrypting parent.

This involved reworking the client side slightly, as the other side of
the plumbing. This means it no longer needs to worry about things like
discarding fully sent buffers.

It also really needs to switch to using a callback instead of a
manifold.stream (probably...I'm still debating that one), but that's for
a different day.

I updated enough of the rest of the related pieces to get the echo test
to pass again (finally!).

Mostly, that just meant converting specs and type hints from ByteBuf to bytes.