Dig back into handshake unit test for messaging
This is almost all just logging and comments, with a few places here and
there where I spotted more places where I was still checking for
booleans left over from ternary sets.

The main change was a spelling error in a keyword that was keeping the
server from responding with the "real" payload.

Now that that part's fixed, the message that wraps the payload it sends
is garbage.

Spotted a problem that was preventing flow control from updating
statistics, but it opens up a can of worms that I'm not willing to
tackle with this commit.