Opened that can of worms
Things are getting into a state where querying for state just times
out.

I really expected to have deadlocks in this area, but there's nothing
else going on in the same timeframe (maybe literally...the bigger-echo
starts, pauses for ~10 seconds, times out, and then times out again when
I try to query for state).

I saw an NPE once. But then I added logging to try to nail it down, and
it went away.