Found 1 bug w/ resending un-ACK'd blocks (WIP)
It was a misplaced parenthesis that was just always sending an else
clause.

There are a *lot* of other changes in here.

My handshake unit test fails with what looks like the deadlock that I've
been expecting. So I've been mucking around with that, mostly seeing if
I could dodge the issue by messing with the executor, buffer sizes, and
retries.

So far, I haven't had any luck.