Work around deadlock in last few commits
By sticking the actual call to ->parent into its own
manifold.deferred/future.

This feels suspiciously like I'm just sweeping it under the rug, but
logic and experience insist that adding another thread to the mix really
is the key, since the problem almost definitely stems from blocking the
"main" event loop.

In practice, this at least relieves the deadlock for my simple handshake
test, so far. So it's good enough for now.

Now I'm back to sending gibberish to the client instead of "the chzbrgr"