Prune back scheduling logging again
Just do that part at the end of the scheduling algorithm, rather than
incrementally as the values are built.

This seems to shave off another order of magnitude of processing time or
so. It's getting down to the point where it might be acceptable.

Debug issues with child monitor thread. For my echo test, at least, it
seems to just be blocking forever.