Tracked down where I broke message passing (WIP)
I originally planned to have a from-parent PipedOutputStream to which
parent-> would write.

That idea went away, because it doesn't make a lot of sense.

The code that I swapped out to actually do that writing is still in
place. So it's been triggering a NPE that I haven't noticed before.