Fix several bugs identified by "big-message" test
The "server" side was receiving a 1K message packet, then the
child-monitor loop was exiting.

That was because the EOF check was wrong: it was triggering on any
truthy value (including the happy path of returning a byte array).

Then I had issues where I was trying to read too many bytes into byte
buffers. Both from a PipedInputStream and when I tried to copy the
subset of my receive work buffer into the destination array that was
actually returning.

And now it fails at the end due to EOF handling.

Well, it might be failing for other reasons also. But that one's the
most obvious at the moment.