Reduce from-child hoop count (1st draft)
Trimmed out a couple of extra procedural convolutions. And uncover a
major hole is applying back-pressure.

This hole was there before, but it wasn't as obvious.

Also eliminated the ::length field from newly created message block
descriptions. It's always been completely redundant.

It's still part of the spec, and I'm sure I have more code downstream
that depends on it. I planned to eliminate all those pieces together in
their own branch.

But retaining it at this point would have been even more of a hack. So
I'm doubling up.