Get echo test working again (more or less)
There were several problems that really all stemmed from code that
needed to be updated to cope with an EOF keyword instead of the original
byte arrays that I originally assumed.

Something goes wrong at the end. I see log messages about failed
attempts to buffer some outgoing message. So, technically, this is a
false positive.

But it's forward progress.