Add EOF signals to echo test (WIP)
Obvious problem with previous commit was that I was trying to write the
EOF signal to a stream that had already been closed. Test for that and
short-circuit if/when it happens.

More insidious problem was that I wasn't actually sending any EOF
markers.

Started doing so. Sending it from the "parent callback" (which is really
a mockup of the "other" child) worked OK. Sending it from the echo test
child-cb broke the test again.

At least the breakage is further along the path.