Fix child's EOF nil message buffering (WIP)
I'm not sure I *like* this fix, but it was a fairly minor change to
switch to a 0-length buffer instead. Which is probably the correct thing
to be sending in this case anyway.

Also, made the message-id an optional field for a ::spec/block
description. We simply do not know it until the block's ready to
send. Even then, it's subject to change if/when we need to resend.

It still doesn't quite get the basic echo test passing again. The
child's EOF message gets stuck in un-sent-blocks before the test
ends.