Track down some problems w/ echo test EOF
Biggest one was that it was waiting for the EOF signal from parent to
send an EOF response, which didn't mesh up with what I thought I was
seeing last time.

I think I've isolated the problem: it looks like the blocks are getting
stuck waiting for an ACK before I send more.