Make echo-test at least look sane (WIP)
The EOF message from the child is getting closer to the parent.

Actually, it has to be reaching the parent, or I'd still be getting
timeout errors.

But it's still showing up in un-sent-blocks instead of being moved over
to un-ackd-blocks the way I expect.

The main change here was adjusting the "global" state to indicate that
we've switched to :send-eof mode to try to get the messages flushed ASAP.