Diagnose obvious issue w/ handshake test: extras
I'm getting back a stream of bytes instead of the original distinct
messages. Because of the way the timing is working out at the moment,
the client receives the first part of the chzbrgr attached to the end of
the :kk that ACKs the client's request for that chzbrgr.

This is annoying. But, honestly, unavoidable.