Refactor the "child" callbacks from handshake test
Really just trimming it down a bit to try to make it a bit more
readable.

Fixed one bug that involved trying to read a keyword as a string. Now I
get a different stack trace at the end of the test. Along with the same
test failures that have been part of the past couple of commits.