Add testing-task dependency on gloss
Because the streaming serialization test on which I'm currently working
has a problem for which it was specifically designed. At least, it's
the top example of how to use aleph.tcp.

Bumped some other library dependency versions while I was at it.

And added some comments about getting rid of my dependencies on logging libraries.