Fixed discrepancies: alt message ioloop scheduler
I'm not sure I've found them all, but I'm pretty certain this is an
improvement that should be merged into master, no matter what.

It turned up while I was working on a different feature branch and seems
like a big enough deal that I didn't want to risk it getting buried in that.