Fix race condition in previous commit (WIP)
My plan for using consume-async to process the actual messages at the
child created a race condition. The ACK for the chzbrgr request and the
chzbrgr arrive in the same message block. They're getting triggered
sequentially, but the ACK handler doesn't have time to update the state
atom before the chzbrgr handler derefs it.

The test was passing with a false positive: there's a state query
request that times out. Added another check to break that.