Start debugging the previous false positive (WIP)
This is mostly just debugging and notes as I start combing through the
logs.

I think the problem stems from not having finished what happens after
the child sends EOF.

Well, the first problem was just marking the test done as soon as we
received the chzbrgr. But that's a symptom.