Start implementing EOF mesages over network (WIP)
Start handling the EOF special cases in from_parent when the other side
sends it.

When one child receives EOF from the other, close its pipe (in
to_child).

Refactored a big inline function inside a reduce just above that to the
top level, just to make things a bit easier to read.