Add decoding layer to server handshake test (WIP)
Basically the same idea as what was added recently to the client portion
of that test, for the gloss communications layer.

Fixed a log entry that I broke when I tried to cast a nanosecond
timestamp into an int by casting it to a long instead.

Messages flow back and forth again, up to the EOF signal.