Wait last unsent before set processed flag (WIP)
It won't send any more new messages once the send-eof-processed flag is
set. So I can't set it until after the last unsent message has been
moved to the unackd queue.

There were a lot of fits and starts getting here, trying to decide where
this should actually happen. I don't like doing this at all, but it's
pretty faithful to the original, and the handshake test is getting
closer to finishing again.