Make caller responsible for stopping main ioloop
Mainly because my handshake test queries for state after everything else
is finished. And that sort of post-mortem analysis seems reasonable to
me.

This version almost looks like that test is really close to passing. The
only obvious failure now is that the client receives more messages than
expected.

I could adjust that easily, but it would be a false positive.

If you pay attention to the logs, there are still issues.