Pay attention to "other" unit tests (WIP)
TL;DR: These are a mess

I really only meant to mess with the ones in message-test. But the
basic-echo test is failing in fun and unusual ways. So I went off on a
tangent to see if I could quickly fix whatever others I've broken with
all the functionality changes in this branch.

Some were easy. I'm punting on the rest for tonight.