Wiggle my way out of to-child parent-monitor log
...corner in which I'd painted myself.

Instead of just writing bytes to a pipe and having a loop waiting for
them to arrive, write a trigger to a stream that has an attached
consumer that will read the bytes.

This isn't fully fleshed out, by any means. But I think the basic idea
should work fairly well.