Start testing last night's changes (WIP)
I figured the echo test would be faster/easier to get running/fixed than
the handshake-test. So I started with it.

This led to starting to convert more of the message ns.

Well, I'd started. I'd just swapped the logging ns alias. Bring back the
original so I can do this incrementally.