Fix some silly problems w/ previous commit (WIP)
I introduced one issue during it, trying to assoc a useful "logging
context" with a boolean "detail" parameter.

That made the test fail almost immediately.

Then there was a TODO comment that I implemented some time ago.

And then my check for whether the EOF had been "processed" for sending
to the other side was based around a boolean instead of the actual
keyword set.

I had high hopes for that one.

This version of the test really has the server receiving the client's
EOF, sending back an ACK, and then just silently dropping the ball until
the test logs out.

At least the failure's a bit more visible