Refactor a bit more of from-child logging (WIP)
Biggest piece, really, was finally refactoring the byte-consumer
function out to a top-level instead of defining it inline inside the
byte-consumer-builder.

I also added a deferred to the top-level from-child caller so it can
flush logs when messages get posted to the main ioloop.

And updated logging/synchronize so there's an entry about the changes in
both. It still isn't great, but at least now the one that had its
lamport clock jump has a record about why.