Finish first draft: revamp from-child logging WIP
Extended that logging with the idea of nested contexts to describe the
purpose of each log (and its forks...which is an idea I also slapped
together without testing).

Converted handshake test to start taking advantage of that multiple
context idea.

This approach builds, and I've started debugging it. But it fails basic
unit testing.