Drag in some numbers from scheduling mismatches
I think this is completely off-topic for the rest of what I'm doing. And
possibly pointless, because I see the same scenario scattered all over
my logs.

But, just in case, it seemed worth capturing a couple of examples that
show where this goes off the rails when I can shift my focus back to
that sort of thing.