Get handshake test working again (WIP)
After converting from-child to new logging approach.

There's a cold hard fact behind this taking so long:
This approach is sadly error-prone. And finicky errors can be tricky to
nail down.

That's true with all logging, of course. But I have to wonder whether
this isn't causing more trouble than it's worth.

On the other hand, the more "normal" approaches cause a *ton* of problems.