Debug most recent problem w/ handshake test WIP
The from-child/byte-consumer callback synchronizes logs between its
builder (which is really the mini ioloop in from-child) and its
caller (which is really the main ioloop).

I was adding an extra ::log/state nesting under itself.

Added a couple of specs to to-child, and fixed another.

Fixed a bunch of log labels in the top-level message ns.

And improved some of the logging details in that handshake test.