Fix problem with shifting units for recent (WIP)
I'd switched it to milliseconds instead of nanoseconds in one spot. This
was wreaking havoc.

This gets my handshake test passing again, but it's still a false
positive. I have something wrong in the flow control logic that sets up
the RTT averages. It's throwing an exception, so they never really get updated.