Update action-trigger to keep log-state in an atom
It was the easiest way I could think of to get its lamport clock to keep
from resetting every time it gets called.

This led to some really garbled log messages in the REPL. The handshake
test broke, but I really couldn't decipher the problem. So I converted
the StdOutLogger to log using an agent.

This fixed both problems.