Convert message.helpers/flag-acked-blocks logging
...to new scheme.

Refactor-renamed it to flag-ackd-blocks for the sake of consistency.

Refactor its acc parameter (which is really a holdover from when it was
an inline function inside a reduce) to state. Again, for consistency.

Added some notes to logging about tempting "improvements" that don't
belong there. (At least, not directly).

Started messing around with from-parent-test, which actually calls into
the other code with which I was messing. It's managed to get even more
broken than I remembered.

I'm really embarassed about the state of these unit tests.