Step back and check rest of unit tests (WIP)
At least the ones that swirl around messaging.

The big-message test was really broken when I tried to log a ByteBuf
after it had been released.

Made the logger handle that sort of thing more gracefully, which got the
test passing again (with error log messages about the problems).

Started trying to track down everywhere I'm logging that so I can just
exclude it.

That approach leaves a lot to be desired. I really want to log the
values associated with that buf.

And it managed to break the test again, with a 1-off error. Or maybe
that was something totally different like a race condition that I just
haven't run across before.