Bump to latest netty version (WIP)
At least it got rid of the annoying startup error that involved a
ClassNotFoundException and jdk.internal.misc.Unsafe.

c.f. https://github.com/netty/netty/issues/6855.

This looked like it was at least in the general vicinity of the core
dump that's coming from trying to build-actual-hello-packet.

As predicted in that ticket, this really didn't help my actual
problem. But it's been something for which I've been itching to find an
excuse for quite a while now.

No idea whether anything else got broken in the process.